Implement bud-06
Add a HEAD /upload endpoint that allows a client to query the server if uploading a file
with certain properties will be possible. This avoids the need of the client having to
call PUT /upload to find out if an upload is possible.

Right now there is not much permission logic build so this PR only adds some basic
HTTP header checks to see if the request is valid. Permission logic will come in a
future PR.